pipelines/training/requirements.txt
transformers>=4.20.0
torch>=1.12.0
peft>=0.4.0
pyyaml>=5.4.0
wandb>=0.13.0
mlflow>=1.26.0
